move data-table component
